Crate resize

Simple resampling library in pure Rust.

Examples

extern crate resize;
use resize::Pixel::Gray8;
use resize::Type::Triangle;
let mut src = vec![0;w1*h1];
let mut dst = vec![0;w2*h2];
let mut resizer = resize::new(w1, h1, w2, h2, Gray8, Triangle);
resizer.resize(&src, &mut dst);

Structs

Filter
Resampling filter.

Resizer
Resampler with preallocated buffers and coeffecients for the given dimensions and filter type.

Enums

Pixel
Supported pixel formats.

Type
Resizing type to use.

Functions

new
Create a new resizer instance. Alias for Resizer::new.

resize
Resize image data to the new dimension in a single step.