Struct resize::Resizer

pub struct Resizer<Pixel: PixelFormat> { /* fields omitted */ }

Resampler with preallocated buffers and coeffecients for the given dimensions and filter type.

Methods

impl<Pixel: PixelFormat> Resizer<Pixel>

fn new(
source_width: usize,
source_heigth: usize,
dest_width: usize,
dest_height: usize,
pixel_format: Pixel,
filter_type: Type
) -> Self

Create a new resizer instance.

fn resize(&mut self, src: &[Pixel::Subpixel], dst: &mut [Pixel::Subpixel])

Resize src image data into dst.

fn resize_stride(
&mut self,
src: &[Pixel::Subpixel],
src_stride: usize,
dst: &mut [Pixel::Subpixel]
)

Resize src image data into dst, skipping stride pixels each row.